+++ lib/clsx/phlex/helper.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Clsx
+  module Phlex
+    module Helper
+      def clsx(*)
+        CLSX.clsx(*)
+      end
+    end
+  end
+end